split influxdb class
