+++ main.sh
@@ -1,2 +1,2 @@
-export PYTHONPATH=./web/:$PYTHONPATH~
+export PYTHONPATH=./back/:./web/:$PYTHONPATH
 gunicorn --bind 0.0.0.0:8080 web.app:app
